Split storage clients into separate files
