+++ activestorage/lib/active_storage/service/azure_storage_service.rb
@@ -4,72 +4,21 @@ gem "azure-storage-blob", "~> 2.0"
 
 require "active_support/core_ext/numeric/bytes"
 require "azure/storage/blob"
-require "azure/storage/blob/default"
-require "azure/storage/common/core/auth/shared_access_signature"
 
 module ActiveStorage
   # Wraps the Microsoft Azure Storage Blob Service as an Active Storage service.
   # See ActiveStorage::Service for the generic API documentation that applies to all services.
   class Service::AzureStorageService < Service
-    # Client that authenticates to Azure Storage via Azure Active Directory
-    # See https://docs.microsoft.com/en-us/azure/storage/common/storage-auth-aad
-    class AzureActiveDirectoryClient
-      attr_reader :storage_account_name, :options, :auth_context, :client_credential
-      attr_accessor :user_delegation_key
-
-      def initialize(storage_account_name, **options)
-        require "adal"
-        gem "adal", "~> 1.0"
-
-        tenant_id = options.delete(:tenant_id) || ""
-        client_id = options.delete(:client_id) || ""
-        client_secret = options.delete(:client_secret) || ""
-
-        if tenant_id.empty? || client_id.empty? || client_secret.empty?
-          raise ArgumentError, "all of tenant_id, client_id, and client_secret must be provided"
-        end
-
-        @storage_account_name = storage_account_name
-        @options = options
-        @auth_context = ADAL::AuthenticationContext.new("login.microsoftonline.com", tenant_id)
-        @client_credential = ADAL::ClientCredential.new(client_id, client_secret)
-        @user_delegation_key = nil
-      end
-
-      def blob_service
-        token = @auth_context.acquire_token_for_client("https://storage.azure.com/", @client_credential)
-        token_credential = Azure::Storage::Common::Core::TokenCredential.new token.access_token
-        token_signer = Azure::Storage::Common::Core::Auth::TokenSigner.new token_credential
-        client = Azure::Storage::Common::Client.create(storage_account_name: @storage_account_name, signer: token_signer)
-        Azure::Storage::Blob::BlobService.new(client: client, api_version: "2018-11-09", **@options)
-      end
-
-      def shared_access_signature
-        if @user_delegation_key.nil? || @user_delegation_key.signed_expiry.to_datetime <= DateTime.now
-          now = Time.now
-          @user_delegation_key = blob_service.get_user_delegation_key(now - 5.minutes, now + 6.days)
-        end
-        Azure::Storage::Common::Core::Auth::SharedAccessSignature.new(@storage_account_name, "", @user_delegation_key)
-      end
-    end
-
-    # Client that authenticates to Azure Storage via account access keys
-    # See https://docs.microsoft.com/en-us/rest/api/storageservices/authorize-with-shared-key
-    class AccessKeyClient
-      attr_reader :blob_service, :shared_access_signature
-
-      def initialize(storage_account_name, storage_access_key, **options)
-        @blob_service = Azure::Storage::Blob::BlobService.create(storage_account_name: storage_account_name, storage_access_key: storage_access_key, **options)
-        @shared_access_signature = Azure::Storage::Common::Core::Auth::SharedAccessSignature.new(storage_account_name, storage_access_key)
-      end
-    end
-
     attr_reader :clients, :container
 
     def initialize(storage_account_name:, storage_access_key: nil, container:, public: false, **options)
-      @clients = storage_access_key.nil? || storage_access_key.empty? \
-        ? AzureActiveDirectoryClient.new(storage_account_name, **options) \
-        : AccessKeyClient.new(storage_account_name, storage_access_key, **options)
+      if storage_access_key.nil? || storage_access_key.empty?
+        require_relative "./azure_storage_service/active_directory_client"
+        @clients = ActiveDirectoryClient.new(storage_account_name, **options)
+      else
+        require_relative "./azure_storage_service/access_key_client"
+        @clients = AccessKeyClient.new(storage_account_name, storage_access_key, **options)
+      end
       @container = container
       @public = public
     end
